requirements/common.txt
# Core Stuff
# -------------------------------------
Django==3.2.12

# Configuration
# -------------------------------------
argon2-cffi==21.3.0
django-environ==0.8.1
django-sites==0.11
python-dotenv==0.20.0
django-cors-headers==3.10.1


# Extensions
# -------------------------------------
pytz==2022.1

# Models
# -------------------------------------
psycopg2-binary==2.9.3

Pillow==8.4.0
django-extensions==3.1.5
django-uuid-upload-path==1.0.0
django-versatileimagefield==2.2

# REST APIs
# -------------------------------------
djangorestframework==3.13.1
django-rest-swagger==2.2.0

# LOGGING
# -------------------------------------
django-log-request-id==2.0.0

# Documentation
# -------------------------------------
mkdocs-material==8.2.8
markdown-include==0.6.0

# Auth Stuff
# -------------------------------------
PyJWT==2.3.0
django-rest-knox==4.2.0
django-mail-templated==2.6.5


# Google Auth
# -------------------------------------
google-api-python-client==2.42.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
